--- Sources/Login/AuthenticationServicing.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol AuthenticationService {
-    func start() async throws
-    @MainActor func handleUniversalLink(_ url: URL) throws
-}
