+++ Sources/Login/LoginCoordinator.swift
@@ -111,7 +111,7 @@ final class LoginCoordinator: NSObject,
     }
     
     private func triggerAuthFlow() async throws {
-        try await authService.start()
+        try await authService.startWebSession()
         guard sessionManager.isReturningUser else {
             launchEnrolmentCoordinator()
             return
