+++ Tests/UnitTests/Tabs/TabManagerCoordinatorTests.swift
@@ -160,21 +160,6 @@ extension TabManagerCoordinatorTests {
     }
     
     @MainActor
-    func test_didRegainFocus_fromLoginCoordinator_withBearerToken() throws {
-        // GIVEN the user has an active session
-//        let loginCoordinator = LoginCoordinator(appWindow: window,
-//                                                root: UINavigationController(),
-//                                                analyticsCenter: mockAnalyticsCenter,
-//                                                sessionManager: mockSessionManager,
-//                                                networkMonitor: MockNetworkMonitor(),
-//                                                isExpiredUser: true)
-        // WHEN the TabManagerCoordinator didRegainFocus from the LoginCoordinator
-//        sut.didRegainFocus(fromChild: loginCoordinator)
-        // THEN no coordinator should be launched
-        XCTAssertEqual(sut.childCoordinators.count, 0)
-    }
-    
-    @MainActor
     func test_performChildCleanup_fromProfileCoordinator_succeeds() async throws {
         let exp = XCTNSNotificationExpectation(
             name: .didLogout,
